requirements.txt
django-discover-runner
coverage
python-dateutil
mock
